More type casts to please MSVC warnings.

+++ src/geodesy/gpsgridder.c
@@ -121,7 +121,7 @@ enum Gpsgridded_enum {	/* Indices for coeff array for normalization */
 GMT_LOCAL void gpsgridder_set_filename (char *name, unsigned int k, unsigned int width, unsigned int mode, unsigned int comp, char *file) {
 	/* Turn name, eigenvalue number k, precision width, mode and comp into a filename, e.g.,
 	 * ("solution.grd", 33, 3, GMT_SVD_INCREMENTAL, 1, file) will give solution_v_inc_033.grd */
-	unsigned int s = strlen (name) - 1;
+	unsigned int s = (unsigned int)strlen (name) - 1;
 	static char *type[3] = {"", "inc", "cum"}, *uv[2] = {"u", "v"};
 	char tmp[GMT_LEN256] = {""};
 	while (name[s] != '.') s--;	/* Wind backwards to start of extension */
